Remove useless code block in gene_relations.py
The code block seems to want to skip to the next gene pair when the
current gene pair's gene ids match using a dictionary lookup. However,
since the continue statement is nested in another for loop, this does
not cause the skip to the next gene pair, and this entire code block is
therefore dead code.