Remove redundant parameters for dep_path()
The dep_path() function is only ever called once, and the values of end1
and end2 when called are equal to start1+1 and start2+1. However, inside
the function itself, the code only ever uses end1-1 and end2-2, which
are the same as the start parameters.

Therefore the two redundant end parameters have been removed.